// components/catalogs/front/src/components/ProductSelection/models/Operator.ts
export enum Operator {
    IS_EMPTY = 'EMPTY',
    IS_NOT_EMPTY = 'NOT EMPTY',
    IN_LIST = 'IN',
    NOT_IN_LIST = 'NOT IN',
    STARTS_WITH = 'STARTS WITH',
    CONTAINS = 'CONTAINS',
    DOES_NOT_CONTAIN = 'DOES NOT CONTAIN',
    EQUALS = '=',
    NOT_EQUAL = '!=',
    IN_CHILDREN_LIST = 'IN CHILDREN',
    NOT_IN_CHILDREN_LIST = 'NOT IN CHILDREN',
    UNCLASSIFIED = 'UNCLASSIFIED',
    IN_LIST_OR_UNCLASSIFIED = 'IN OR UNCLASSIFIED',
    LOWER_THAN = '<',
    LOWER_OR_EQUAL_THAN = '<=',
    GREATER_THAN = '>',
    GREATER_OR_EQUAL_THAN = '>=',
    BETWEEN = 'BETWEEN',
    NOT_BETWEEN = 'NOT BETWEEN',
}